Add DB/Influx2cli
Use the Influx2 CLI to manage Influx2 buckets.  The Influx2 server can
be hosted in a local machine, in the datacenter, or on the Influx2
cloud.

Note that the config file `~/.influxdbv2/conf` is synced across all
development machines.

To interact with an InfluxDB instance, you will need:
- a token
- a URL
- an organization
- a bucket

These can be setup in the InfluxUI, on the cloud or a local machine.